split file and refactor theme
